refactor(handlers): use common summary api stuct methods; rename vulnerabilityreport package to trivyreport package
