+++ internal/handlers/trivyreport/doc.go
@@ -1,0 +1,2 @@
+// Package trivyreport provides utility functions for converting Trivy reports into the SBOMScanner VulnerabilityReport format.
+package trivyreport
