--- internal/handlers/vulnerabilityreport/doc.go
@@ -1,3 +1,0 @@
-// Package vulnerabilityreport provides functions to convert from other
-// tools formats, into the sbomscanner format.
-package vulnerabilityreport
