--- internal/handlers/vulnerabilityreport/summary.go
@@ -1,39 +1,0 @@
-package vulnerabilityreport
-
-import (
-	storagev1alpha1 "github.com/kubewarden/sbomscanner/api/storage/v1alpha1"
-)
-
-func ComputeSummary(results []storagev1alpha1.Result) storagev1alpha1.Summary {
-	var critical, high, medium, low, unknown, suppressed int
-
-	for _, result := range results {
-		for _, vuln := range result.Vulnerabilities {
-			if vuln.Suppressed {
-				suppressed++
-				continue
-			}
-			switch vuln.Severity {
-			case "CRITICAL":
-				critical++
-			case "HIGH":
-				high++
-			case "MEDIUM":
-				medium++
-			case "LOW":
-				low++
-			case "UNKNOWN":
-				unknown++
-			}
-		}
-	}
-
-	return storagev1alpha1.Summary{
-		Critical:   critical,
-		High:       high,
-		Medium:     medium,
-		Low:        low,
-		Unknown:    unknown,
-		Suppressed: suppressed,
-	}
-}
