--- internal/handlers/vulnerabilityreport/summary_test.go
@@ -1,43 +1,0 @@
-package vulnerabilityreport
-
-import (
-	"testing"
-
-	"github.com/stretchr/testify/assert"
-
-	storagev1alpha1 "github.com/kubewarden/sbomscanner/api/storage/v1alpha1"
-)
-
-func TestComputeSummary(t *testing.T) {
-	results := []storagev1alpha1.Result{
-		{
-			Vulnerabilities: []storagev1alpha1.Vulnerability{
-				{Severity: "CRITICAL", Suppressed: false},
-				{Severity: "HIGH", Suppressed: false},
-				{Severity: "MEDIUM", Suppressed: false},
-				{Severity: "LOW", Suppressed: false},
-				{Severity: "UNKNOWN", Suppressed: false},
-				{Severity: "HIGH", Suppressed: true}, // suppressed, shouldn't count in HIGH
-			},
-		},
-		{
-			Vulnerabilities: []storagev1alpha1.Vulnerability{
-				{Severity: "CRITICAL", Suppressed: false},
-				{Severity: "MEDIUM", Suppressed: true}, // another suppressed
-			},
-		},
-	}
-
-	summary := ComputeSummary(results)
-
-	expected := storagev1alpha1.Summary{
-		Critical:   2,
-		High:       1,
-		Medium:     1,
-		Low:        1,
-		Unknown:    1,
-		Suppressed: 2,
-	}
-
-	assert.Equal(t, expected, summary)
-}
